+++ packages/@ember/-internals/glimmer/lib/modifiers/action.ts
@@ -1,4 +1,4 @@
-import { Owner } from '@ember/-internals/owner';
+import type { Owner } from '@ember/-internals/owner';
 import { uuid } from '@ember/-internals/utils';
 import { ActionManager, EventDispatcher, isSimpleClick } from '@ember/-internals/views';
 import { assert } from '@ember/debug';
@@ -6,7 +6,7 @@ import { flaggedInstrument } from '@ember/instrumentation';
 import { join } from '@ember/runloop';
 import { registerDestructor } from '@glimmer/destroyable';
 import { DEBUG } from '@glimmer/env';
-import {
+import type {
   CapturedArguments,
   CapturedNamedArguments,
   CapturedPositionalArguments,
@@ -14,8 +14,9 @@ import {
 } from '@glimmer/interfaces';
 import { setInternalModifierManager } from '@glimmer/manager';
 import { isInvokableRef, updateRef, valueForRef } from '@glimmer/reference';
-import { createUpdatableTag, UpdatableTag } from '@glimmer/validator';
-import { SimpleElement } from '@simple-dom/interface';
+import type { UpdatableTag } from '@glimmer/validator';
+import { createUpdatableTag } from '@glimmer/validator';
+import type { SimpleElement } from '@simple-dom/interface';
 
 const MODIFIERS = ['alt', 'shift', 'meta', 'ctrl'];
 const POINTER_EVENT_TYPE_REGEX = /^click|mouse|touch/;
@@ -50,7 +51,7 @@ export let ActionHelper = {
   registerAction(actionState: ActionState) {
     let { actionId } = actionState;
 
-    (ActionManager.registeredActions as Record<string, unknown>)[actionId] = actionState;
+    ActionManager.registeredActions[actionId] = actionState;
 
     return actionId;
   },
@@ -58,7 +59,7 @@ export let ActionHelper = {
   unregisterAction(actionState: ActionState) {
     let { actionId } = actionState;
 
-    delete (ActionManager.registeredActions as Record<string, unknown>)[actionId];
+    delete ActionManager.registeredActions[actionId];
   },
 };