+++ packages/@ember/-internals/views/lib/views/core_view.ts
@@ -1,6 +1,7 @@
-import { Renderer, View } from '@ember/-internals/glimmer/lib/renderer';
+import type { Renderer, View } from '@ember/-internals/glimmer/lib/renderer';
 import { inject } from '@ember/-internals/metal';
 import { ActionHandler, Evented, FrameworkObject } from '@ember/-internals/runtime';
+import type { ViewState } from './states';
 import states from './states';
 
 /**
@@ -26,8 +27,8 @@ class CoreView extends FrameworkObject.extend(Evented, ActionHandler) {
 
   declare _states: typeof states;
 
-  _state: unknown;
-  _currentState: unknown;
+  declare _state: keyof typeof states;
+  declare _currentState: ViewState;
 
   _superTrigger?: Evented['trigger'];
   _superHas?: Evented['has'];