+++ packages/@ember/-internals/views/lib/views/states.ts
@@ -2,10 +2,11 @@ import preRender from './states/pre_render';
 import hasElement from './states/has_element';
 import inDOM from './states/in_dom';
 import destroying from './states/destroying';
-import Component from '@ember/component';
+import type Component from '@ember/component';
 
 export interface ViewState {
   enter?(view: Component): void;
+  exit?(view: Component): void;
   appendChild(): void;
   handleEvent(view: Component, eventName: string, event: Event): boolean;
   rerender(view: Component): void;